refactor(options): split production Stitch controllers
